backend/requirements.txt
Flask
Flask-Cors
pandas
numpy
scikit-learn
pymongo
bcrypt
PyJWT
